move instancetooltip and flatten details folder
