+++ airflow/www/static/js/dag/details/ConfirmDialog.tsx
