+++ airflow/www/static/js/dag/details/taskInstance/Details.tsx
