+++ airflow/www/static/js/dag/details/taskInstance/Logs/utils.ts
