+++ airflow/www/static/js/dag/details/taskInstance/taskActions/MarkFailed.jsx
@@ -26,17 +26,16 @@ import {
 } from '@chakra-ui/react';
 
 import ActionButton from './ActionButton';
+import { useConfirmMarkTask, useMarkFailedTask } from '../../../../api';
 import ConfirmDialog from '../../ConfirmDialog';
-import { useClearTask } from '../../../../../api';
-import { getMetaValue } from '../../../../../utils';
+import { getMetaValue } from '../../../../utils';
 
 const canEdit = getMetaValue('can_edit') === 'True';
 
-const Run = ({
+const MarkFailed = ({
   dagId,
   runId,
   taskId,
-  executionDate,
   mapIndexes,
 }) => {
   const [affectedTasks, setAffectedTasks] = useState([]);
@@ -51,31 +50,27 @@ const Run = ({
   const [upstream, setUpstream] = useState(false);
   const onToggleUpstream = () => setUpstream(!upstream);
 
-  const [downstream, setDownstream] = useState(true);
+  const [downstream, setDownstream] = useState(false);
   const onToggleDownstream = () => setDownstream(!downstream);
-
-  const [recursive, setRecursive] = useState(true);
-  const onToggleRecursive = () => setRecursive(!recursive);
-
-  const [failed, setFailed] = useState(false);
-  const onToggleFailed = () => setFailed(!failed);
 
   // Confirm dialog open/close
   const { isOpen, onOpen, onClose } = useDisclosure();
 
-  const { mutateAsync: clearTask, isLoading } = useClearTask({
-    dagId, runId, taskId, executionDate,
+  const { mutateAsync: markFailedMutation, isLoading: isMarkLoading } = useMarkFailedTask({
+    dagId, runId, taskId,
+  });
+  const {
+    mutateAsync: confirmChangeMutation, isLoading: isConfirmLoading,
+  } = useConfirmMarkTask({
+    dagId, runId, taskId, state: 'failed',
   });
 
   const onClick = async () => {
-    const data = await clearTask({
+    const data = await confirmChangeMutation({
       past,
       future,
       upstream,
       downstream,
-      recursive,
-      failed,
-      confirmed: false,
       mapIndexes,
     });
     setAffectedTasks(data);
@@ -83,19 +78,18 @@ const Run = ({
   };
 
   const onConfirm = async () => {
-    await clearTask({
+    await markFailedMutation({
       past,
       future,
       upstream,
       downstream,
-      recursive,
-      failed,
-      confirmed: true,
       mapIndexes,
     });
     setAffectedTasks([]);
     onClose();
   };
+
+  const isLoading = isMarkLoading || isConfirmLoading;
 
   return (
     <Flex justifyContent="space-between" width="100%">
@@ -104,28 +98,20 @@ const Run = ({
         <ActionButton bg={future && 'gray.100'} onClick={onToggleFuture} name="Future" />
         <ActionButton bg={upstream && 'gray.100'} onClick={onToggleUpstream} name="Upstream" />
         <ActionButton bg={downstream && 'gray.100'} onClick={onToggleDownstream} name="Downstream" />
-        <ActionButton bg={recursive && 'gray.100'} onClick={onToggleRecursive} name="Recursive" />
-        <ActionButton bg={failed && 'gray.100'} onClick={onToggleFailed} name="Failed" />
       </ButtonGroup>
-      <Button
-        colorScheme="blue"
-        onClick={onClick}
-        isLoading={isLoading}
-        isDisabled={!canEdit}
-        title="Clearing deletes the previous state of the task instance, allowing it to get re-triggered by the scheduler or a backfill command"
-      >
-        Clear
+      <Button colorScheme="red" onClick={onClick} isLoading={isLoading} isDisabled={!canEdit}>
+        Mark Failed
       </Button>
       <ConfirmDialog
         isOpen={isOpen}
         onClose={onClose}
         onConfirm={onConfirm}
         isLoading={isLoading}
-        description="Task instances you are about to clear:"
+        description="Task instances you are about to mark as failed:"
         body={affectedTasks}
       />
     </Flex>
   );
 };
 
-export default Run;
+export default MarkFailed;
